moves a guice module from runtime to common
* Includes refactoring of other inessentials that had crept into the
 ConfigModule class
* Lets refrain from not putting things where they don't belong and
 taking shortcuts that take longer to fix later
