+++ redriver/src/test/java/com/flipkart/flux/redriver/boot/RedriverTestModule.java
@@ -1,0 +1,4 @@
+package com.flipkart.flux.redriver.boot;
+
+public class RedriverTestModule {
+}
