+++ runtime/src/main/java/com/flipkart/flux/initializer/FluxInitializer.java
@@ -16,9 +16,7 @@ package com.flipkart.flux.initializer;
 import com.flipkart.flux.MigrationUtil.MigrationsRunner;
 import com.flipkart.flux.client.FluxClientInterceptorModule;
 import com.flipkart.flux.deploymentunit.ExecutableRegistryPopulator;
-import com.flipkart.flux.guice.module.ConfigModule;
-import com.flipkart.flux.guice.module.ContainerModule;
-import com.flipkart.flux.guice.module.HibernateModule;
+import com.flipkart.flux.guice.module.*;
 import com.flipkart.flux.impl.boot.TaskModule;
 import com.flipkart.polyguice.core.support.Polyguice;
 import org.slf4j.Logger;
@@ -118,6 +116,8 @@ public class FluxInitializer {
                 configModule,
                 new HibernateModule(),
                 new ContainerModule(),
+                new DeploymentUnitModule(),
+                new AkkaModule(),
                 new TaskModule(),
                 new FluxClientInterceptorModule()
         );
